internal: upgrade hostd
