+++ bus/bus.go
@@ -29,6 +29,7 @@ import (
 	"go.sia.tech/renterd/tracing"
 	"go.sia.tech/renterd/wallet"
 	"go.sia.tech/renterd/webhooks"
+	"go.sia.tech/siad/modules"
 	"go.uber.org/zap"
 )
 
@@ -50,10 +51,13 @@ func NewClient(addr, password string) *Client {
 type (
 	// A ChainManager manages blockchain state.
 	ChainManager interface {
-		AcceptBlock(context.Context, types.Block) error
+		AcceptBlock(types.Block) error
+		BlockAtHeight(height uint64) (types.Block, bool)
+		IndexAtHeight(height uint64) (types.ChainIndex, error)
 		LastBlockTime() time.Time
-		Synced(ctx context.Context) bool
-		TipState(ctx context.Context) consensus.State
+		Subscribe(s modules.ConsensusSetSubscriber, ccID modules.ConsensusChangeID, cancel <-chan struct{}) error
+		Synced() bool
+		TipState() consensus.State
 	}
 
 	// A Syncer can connect to other peers and synchronize the blockchain.
@@ -66,10 +70,13 @@ type (
 
 	// A TransactionPool can validate and relay unconfirmed transactions.
 	TransactionPool interface {
+		AcceptTransactionSet(txns []types.Transaction) error
 		RecommendedFee() types.Currency
+		Subscribe(subscriber modules.TransactionPoolSubscriber)
 		Transactions() []types.Transaction
-		AddTransactionSet(txns []types.Transaction) error
 		UnconfirmedParents(txn types.Transaction) ([]types.Transaction, error)
+
+		Close() error
 	}
 
 	// A Wallet can spend and receive siacoins.
@@ -217,7 +224,7 @@ func (b *bus) consensusAcceptBlock(jc jape.Context) {
 	if jc.Decode(&block) != nil {
 		return
 	}
-	if jc.Check("failed to accept block", b.cm.AcceptBlock(jc.Request.Context(), block)) != nil {
+	if jc.Check("failed to accept block", b.cm.AcceptBlock(block)) != nil {
 		return
 	}
 }
@@ -242,12 +249,12 @@ func (b *bus) syncerConnectHandler(jc jape.Context) {
 }
 
 func (b *bus) consensusStateHandler(jc jape.Context) {
-	jc.Encode(b.consensusState(jc.Request.Context()))
+	jc.Encode(b.consensusState())
 }
 
 func (b *bus) consensusNetworkHandler(jc jape.Context) {
 	jc.Encode(api.ConsensusNetwork{
-		Name: b.cm.TipState(jc.Request.Context()).Network.Name,
+		Name: b.cm.TipState().Network.Name,
 	})
 }
 
@@ -263,7 +270,7 @@ func (b *bus) txpoolTransactionsHandler(jc jape.Context) {
 func (b *bus) txpoolBroadcastHandler(jc jape.Context) {
 	var txnSet []types.Transaction
 	if jc.Decode(&txnSet) == nil {
-		jc.Check("couldn't broadcast transaction set", b.tp.AddTransactionSet(txnSet))
+		jc.Check("couldn't broadcast transaction set", b.tp.AcceptTransactionSet(txnSet))
 	}
 }
 
@@ -378,7 +385,7 @@ func (b *bus) walletFundHandler(jc jape.Context) {
 		fee := b.tp.RecommendedFee().Mul64(uint64(encodedLen(txn)))
 		txn.MinerFees = []types.Currency{fee}
 	}
-	toSign, err := b.w.FundTransaction(b.cm.TipState(jc.Request.Context()), &txn, wfr.Amount.Add(txn.MinerFees[0]), b.tp.Transactions())
+	toSign, err := b.w.FundTransaction(b.cm.TipState(), &txn, wfr.Amount.Add(txn.MinerFees[0]), b.tp.Transactions())
 	if jc.Check("couldn't fund transaction", err) != nil {
 		return
 	}
@@ -399,7 +406,7 @@ func (b *bus) walletSignHandler(jc jape.Context) {
 	if jc.Decode(&wsr) != nil {
 		return
 	}
-	err := b.w.SignTransaction(b.cm.TipState(jc.Request.Context()), &wsr.Transaction, wsr.ToSign, wsr.CoveredFields)
+	err := b.w.SignTransaction(b.cm.TipState(), &wsr.Transaction, wsr.ToSign, wsr.CoveredFields)
 	if jc.Check("couldn't sign transaction", err) == nil {
 		jc.Encode(wsr.Transaction)
 	}
@@ -415,7 +422,7 @@ func (b *bus) walletRedistributeHandler(jc jape.Context) {
 		return
 	}
 
-	cs := b.cm.TipState(jc.Request.Context())
+	cs := b.cm.TipState()
 	txn, toSign, err := b.w.Redistribute(cs, wfr.Outputs, wfr.Amount, b.tp.RecommendedFee(), b.tp.Transactions())
 	if jc.Check("couldn't redistribute money in the wallet into the desired outputs", err) != nil {
 		return
@@ -426,7 +433,7 @@ func (b *bus) walletRedistributeHandler(jc jape.Context) {
 		return
 	}
 
-	if jc.Check("couldn't broadcast the transaction", b.tp.AddTransactionSet([]types.Transaction{txn})) != nil {
+	if jc.Check("couldn't broadcast the transaction", b.tp.AcceptTransactionSet([]types.Transaction{txn})) != nil {
 		b.w.ReleaseInputs(txn)
 		return
 	}
@@ -442,7 +449,6 @@ func (b *bus) walletDiscardHandler(jc jape.Context) {
 }
 
 func (b *bus) walletPrepareFormHandler(jc jape.Context) {
-	ctx := jc.Request.Context()
 	var wpfr api.WalletPrepareFormRequest
 	if jc.Decode(&wpfr) != nil {
 		return
@@ -455,7 +461,7 @@ func (b *bus) walletPrepareFormHandler(jc jape.Context) {
 		jc.Error(errors.New("no renter key provided"), http.StatusBadRequest)
 		return
 	}
-	cs := b.cm.TipState(ctx)
+	cs := b.cm.TipState()
 
 	fc := rhpv2.PrepareContractFormation(wpfr.RenterKey, wpfr.HostKey, wpfr.RenterFunds, wpfr.HostCollateral, wpfr.EndHeight, wpfr.HostSettings, wpfr.RenterAddress)
 	cost := rhpv2.ContractFormationCost(cs, fc, wpfr.HostSettings.ContractPrice)
@@ -490,7 +496,7 @@ func (b *bus) walletPrepareRenewHandler(jc jape.Context) {
 		jc.Error(errors.New("no renter key provided"), http.StatusBadRequest)
 		return
 	}
-	cs := b.cm.TipState(jc.Request.Context())
+	cs := b.cm.TipState()
 
 	// Create the final revision from the provided revision.
 	finalRevision := wprr.Revision
@@ -1439,17 +1445,17 @@ func (b *bus) paramsHandlerUploadGET(jc jape.Context) {
 
 	jc.Encode(api.UploadParams{
 		ContractSet:   contractSet,
-		CurrentHeight: b.cm.TipState(jc.Request.Context()).Index.Height,
+		CurrentHeight: b.cm.TipState().Index.Height,
 		GougingParams: gp,
 		UploadPacking: uploadPacking,
 	})
 }
 
-func (b *bus) consensusState(ctx context.Context) api.ConsensusState {
+func (b *bus) consensusState() api.ConsensusState {
 	return api.ConsensusState{
-		BlockHeight:   b.cm.TipState(ctx).Index.Height,
+		BlockHeight:   b.cm.TipState().Index.Height,
 		LastBlockTime: b.cm.LastBlockTime(),
-		Synced:        b.cm.Synced(ctx),
+		Synced:        b.cm.Synced(),
 	}
 }
 
@@ -1476,7 +1482,7 @@ func (b *bus) gougingParams(ctx context.Context) (api.GougingParams, error) {
 		b.logger.Panicf("failed to unmarshal redundancy settings '%s': %v", rss, err)
 	}
 
-	cs := b.consensusState(ctx)
+	cs := b.consensusState()
 
 	return api.GougingParams{
 		ConsensusState:     cs,
@@ -1688,7 +1694,7 @@ func (b *bus) contractTaxHandlerGET(jc jape.Context) {
 	if jc.DecodeParam("payout", (*api.ParamCurrency)(&payout)) != nil {
 		return
 	}
-	cs := b.cm.TipState(jc.Request.Context())
+	cs := b.cm.TipState()
 	jc.Encode(cs.FileContractTax(types.FileContract{Payout: payout}))
 }
